+++ src/paddlefx/pyeval.py
@@ -21,7 +21,6 @@ from .bytecode_transformation import (
     transform_code_object,
     unique_id,
 )
-from .cache_manager import GuardedCode
 from .codegen import PyCodegen
 from .output_graph import OutputGraph
 from .paddle_utils import TensorType
@@ -34,51 +33,7 @@ from .variables.builder import GraphArg
 
 if TYPE_CHECKING:
     # import opcode
-    from .cache_manager import GuardedCodes
-
-
-class CodeCacheManager:
-    cache_dict: dict[types.CodeType, GuardedCodes] = {}
-
-    @classmethod
-    def add_cache(cls, code: types.CodeType, guarded_code: GuardedCode):
-        cls.cache_dict.setdefault(code, [])
-        cls.cache_dict[code].append(guarded_code)
-
-    @classmethod
-    def get_cache(cls, frame: types.FrameType) -> GuardedCode | None:
-        code: types.CodeType = frame.f_code
-        if code not in cls.cache_dict:
-            print(f"Firstly call {code}\n")
-            return None
-        return cls.lookup(frame, cls.cache_dict[code])
-
-    @classmethod
-    def clear_cache(cls):
-        cls.cache_dict.clear()
-
-    @classmethod
-    def lookup(
-        cls, frame: types.FrameType, guarded_codes: GuardedCodes
-    ) -> GuardedCode | None:
-        for guarded_code in guarded_codes:
-            try:
-                guard_fn = guarded_code.guard_fn
-                if guard_fn(frame):
-                    print(
-                        f"[Cache]: Cache hit, GuardFunction is {guard_fn}\n",
-                    )
-                    return guarded_code
-                else:
-                    print(
-                        f"[Cache]: Cache miss, GuardFunction is {guard_fn}\n",
-                    )
-            except Exception as e:
-                print(f"[Cache]: GuardFunction function error: {e}\n")
-                continue
-
-        print("[Cache]: all guards missed\n")
-        return None
+    pass
 
 
 def tos_op_wrapper(fn: Callable):
